+++ src/js/platform/browser/no_achievements.js
@@ -1,10 +1,6 @@
 import { AchievementsInterface } from "../achievements";
 
 export class NoAchievements extends AchievementsInterface {
-    initialize() {
-        return Promise.resolve();
-    }
-
     load() {
         return Promise.resolve();
     }
